split seq2seq
