+++ src/data_util.py
@@ -62,7 +62,7 @@ def normalizeString(s):
     s = re.sub(r"([.!?])", r" \1", s)
     s = re.sub(r"[^a-zA-Z.!?]+", r" ", s)
     return s
-  
+
 
 def unicodeToAscii(s):
     return ''.join(
